`[SHINUTI]`: Sotsu Omeshiki Full: Increase Oni (660 -> 680)
This one is a little trickier to justify, because this song doesn't have an AC (Nijiro) release, and is only available on NS2/PS4. This means that the wikiwiki.jp page only has the "red text" score, which is *always* too low in practice.

https://wikiwiki.jp/taiko-fumen/%E5%8F%8E%E9%8C%B2%E6%9B%B2/%E3%81%8A%E3%81%AB/%E3%81%9D%E3%81%A4%E3%81%8A%E3%82%81%E3%81%97%E3%81%8D%E3%83%BB%E3%81%B5%E3%82%8B

So, to roughly estimate what the score should be, I use a heuristic:

  - Songs with `n_GOOD ~= 10x n_OK ~= 100x n_BAD` should have a Nijiro score of about 950k.

For Sotsu Omeshiki Full, imagine an accuracy of 1344/130/13. According to the heuristic above, this should have a score of ~950k, if not a little more (since the number of GOOD is above 10 * 130 OK).

- 660 points: 929,940
- 670 points: 943,380
- 680 points: 958,120

660 and 670 produce scores that are too low. So, 680 points makes the most intuitive sense to me based on what I know about other "correct" Nijiro/AC scores.